+++ src/components/header/index.tsx
@@ -1,10 +1,14 @@
 import React from 'react'
 import Image from 'next/image'
 
+import { SidebarMenu } from './sidebar-menu'
+
 export const Header = () => {
   return (
     <header>
-      <div className="flex justify-center p-4">
+      <div className="flex p-4 justify-between shadow-2xl">
+        <SidebarMenu />
+
         <Image
           src="/assets/svg/Header.svg"
           alt="Header"
@@ -13,6 +17,8 @@ export const Header = () => {
           placeholder="blur"
           blurDataURL={'/assets/svg/Header.svg'}
         />
+
+        <span></span>
       </div>
     </header>
   )