+++ CHANGELOG.md
@@ -2,21 +2,45 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### [0.1.5](https://github.com/mokkapps/changelog-generator-demo/compare/v0.1.4...v0.1.5) (2021-09-19)
-
-
-### Bug Fixes
-
-* crlf error ([98c5e42](https://github.com/mokkapps/changelog-generator-demo/commits/98c5e4281c36985fbe7daca749b2827da0739fd0))
-
-### [0.1.4](https://github.com/mokkapps/changelog-generator-demo/compare/v0.1.3...v0.1.4) (2021-09-19)
+### [0.1.7](https://github.com/mokkapps/changelog-generator-demo/compare/v0.1.6...v0.1.7) (2021-09-20)
 
 
 ### Features
 
-* go to selected pages ([ede062f](https://github.com/mokkapps/changelog-generator-demo/commits/ede062f1cd802ed79185b19eb1b0d031d5b2e766))
-* home screen ([2cc3b54](https://github.com/mokkapps/changelog-generator-demo/commits/2cc3b544e7ea964135bebd3d50fbb1911f4e22a4))
-* profile picture component ([ac88f7d](https://github.com/mokkapps/changelog-generator-demo/commits/ac88f7ddb21cc09853a3cb11a369f6a348886b77))
+* logout ([007b232](https://github.com/mokkapps/changelog-generator-demo/commits/007b23222cc5078921ad2e3258a075e39a30db9b))
+* new columns on database ([5a56869](https://github.com/mokkapps/changelog-generator-demo/commits/5a56869bd9f6a4847824366e67b896b5d61094e0))
+* sidebar ([eda875c](https://github.com/mokkapps/changelog-generator-demo/commits/eda875c7bb2d1b14e85602256baa8f29621033c2))
+
+
+### Bug Fixes
+
+* fix secure home ([31ba73a](https://github.com/mokkapps/changelog-generator-demo/commits/31ba73a4d2208bff663ba4d5359dbc8257d70564))
+* logout ([8ae6d78](https://github.com/mokkapps/changelog-generator-demo/commits/8ae6d78703678ce172456532178652d33f1f00f7))
+
+### [0.1.6](https://github.com/mokkapps/changelog-generator-demo/compare/v0.1.5...v0.1.6) (2021-09-20)
+
+### Features
+
+- event class ([eb10baa](https://github.com/mokkapps/changelog-generator-demo/commits/eb10baada0b328a4527edc4512242f46581aa197))
+- make routes on sign-in better ([852d807](https://github.com/mokkapps/changelog-generator-demo/commits/852d807fa0315a7fa32edaac046e95239b348733))
+
+### Bug Fixes
+
+- events on cancel page ([a8daf60](https://github.com/mokkapps/changelog-generator-demo/commits/a8daf60d9709c65e978fec185a6cf4a768f96e01))
+
+### [0.1.5](https://github.com/mokkapps/changelog-generator-demo/compare/v0.1.4...v0.1.5) (2021-09-19)
+
+### Bug Fixes
+
+- crlf error ([98c5e42](https://github.com/mokkapps/changelog-generator-demo/commits/98c5e4281c36985fbe7daca749b2827da0739fd0))
+
+### [0.1.4](https://github.com/mokkapps/changelog-generator-demo/compare/v0.1.3...v0.1.4) (2021-09-19)
+
+### Features
+
+- go to selected pages ([ede062f](https://github.com/mokkapps/changelog-generator-demo/commits/ede062f1cd802ed79185b19eb1b0d031d5b2e766))
+- home screen ([2cc3b54](https://github.com/mokkapps/changelog-generator-demo/commits/2cc3b544e7ea964135bebd3d50fbb1911f4e22a4))
+- profile picture component ([ac88f7d](https://github.com/mokkapps/changelog-generator-demo/commits/ac88f7ddb21cc09853a3cb11a369f6a348886b77))
 
 ### 0.1.3 (2021-09-19)